requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
cachetools==5.5.1
certifi==2023.7.22
charset-normalizer==3.4.1
colorama==0.4.6
diffusers==0.32.2
distro==1.9.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1